Move headArguments.tsx
